postgres_to_es/requirements.txt
psycopg2-binary==2.9.3
pydantic==1.9.0
elasticsearch==8.2.0
